+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add remote Swift packages to the Frameworks build phase https://github.com/tuist/XcodeProj/pull/487 by @kwridan
 - System library added to a group has empty path https://github.com/tuist/XcodeProj/pull/488 by @damirdavletov
+- Fix Products group serialisation with temporary ids https://github.com/tuist/XcodeProj/pull/489 by @damirdavletov
 
 ## 7.1.0